yes
The writing of this article is grounded by observations and simultaneous assessments of the external and internal environments of educational institutions allowing education managers to be able to identify the different types of opportunities available to successfully formulate and implement various educational plans. This comprehensive design can be done through a process of action known as a strategic environmental analysis. Strategic concept is more effective in allocating existing resources in achieving educational goals. The existence of the implementation of Islamic Education will be a means, the media in the distribution of vision and mission in addition to implications on improving the quality of faith, taqwa, even faith (leader) for the devout. Environmental analysis is a stakeholder analysis, a very important instrument for understanding the social and institutional context of a program / project activity. This instrument is used to identify stakeholder interests, interests, and influence on ongoing program / project activities.