Strategi Alternatif Lembaga Pendidikan Islam
There are changes that occur in our education which impact on regional competitive build towards a global competitive, both in the formal institutions of national and Islamic institutions. So this raises complex issues that wraps acceleration of development in promoting an Islamic educational institutions in rural areas. These problems include institutional, leadership, finance, personnel, curriculum and educational model. Each of these components can inhibit the saving potential and within certain limits can threaten the existence, sustainability, competitiveness, and progress of Islamic educational institutions. The question of research: how is each of these institutions are trying to attract views of rural communities by providing an alternative strategy such as the advantages and characteristics of the institution in the decentralization and globalization?. This research resulted in four alternative bid that is modeling the Islamic educational institutions business model in the financing strategy, improve excellence at regular intervals sustainable in the interests of rural communities, implement the task of public relations in spreading the advancement of excellence of the school, and no achievement or excellence will find both nationally and internationally. Thus meaning that the top schools are part of a plan to be achieved can run well and there are no gaps in it and is able to work by well.